fix(chat): add floating overlay on all screens if panels opened (Issue #5376)
